// src/app/shared/interfaces/offer-interface.ts
export interface Offer {
    id?: number;
    originalLgtUrl?: string;
    firstImageUrl?: string;
    destinationCountry?: string;
    destinationDistrict?: string;
    destinationDistrictZone?: string;
    destinationCity?: string;
    destinationCityZone?: string;
    destinationHotelName?: string;
    destinationHotelStars?: string;
    destinationHotelServicesType?: string;
    includedServicesPlaneTickets?: boolean;
}